Phase 6: Modularize CppMcpServer tool handlers
- Move search tool handlers to clang_index_mcp/_mcp/tool_handlers/search_tools.py.
- Move class-hierarchy handler to tool_handlers/hierarchy_tools.py.
- Move call-graph handlers to tool_handlers/call_graph_tools.py.
- Move project/refresh handlers to tool_handlers/project_tools.py.
- Move query policy helpers to tool_handlers/policy_tools.py.
- Move transport/signal helpers to tool_handlers/transport_tools.py.
- Keep cpp_mcp_server.py as a thin orchestrator exposing list_tools, call_tool,
  main, and server lifecycle helpers.
- Full test suite passes: 1432 passed, 17 skipped, 4 xpassed.
